+++ CHANGELOG.md
@@ -1,13 +1,26 @@
 # :arrows_counterclockwise: VMware SRM As Built Report Changelog
 
-## [0.1.0] - YYYY-MM-DD
+## [0.2.0] - 2021-11-01
 
 ### Added
 
+- Added vCenter Summary.
+- Added SRM Summary Information.
+  - Added SRM Licensing information.
+  - Added SRM ACL Permissions information.
+- Added SRM Protection Group Information.
+  - Added Protection Group VR & SAN type Information.
+  - Added Per VM Protection Properties.
+- Added SRM Recovery Plan Information.
+  - Added Per VM Recovery Settings.
+- Added SRM Inventory Mapping Information.
+- Added Protected & Recovery Site Infromation.
+
 ### Changed
 
-### Fixed
+- Implemented remote validation of vCenter credentials (separate credentials for protected and recovery sites).
+- The code was changed to match the new module structure.
 
-### Removed
+## [0.1.0] - 2021-10-16
 
-
+- Initial release from @mattallford.